+++ src/ShippingFile/Views/Cnab400.php
@@ -19,6 +19,8 @@ use aryelgois\BankInterchange;
  */
 class Cnab400 extends BankInterchange\ShippingFile\View
 {
+    const TITLE_LIMIT = 999997;
+
     /**
      * Adds a File Header
      */