+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Column `accept` in `titles` table
 - Columns for `interest`
 - Column `agency_account_cd`
+- ShippingFile View `TITLE_LIMIT`
 
 ### Changed
 - Bump [aryelgois/medools] version